apps/dash-lastodash/requirements.txt
dash==2.9.3
dash-daq==0.5.0
dash-table==5.0.0
plotly==5.15.0
lasio==0.23
pandas==1.1.5
psycopg2-binary==2.8.6
gunicorn
